+++ game/src/renderers/bulletrenderer.cpp
@@ -1,11 +1,10 @@
 #include "renderers/bulletrenderer.h"
 
-void BulletRenderer::render(Gfx &gfx, Position &position, Shape &shape, Direction &, RGB &colour) const
-{
-    gfx.drawRect(position.X() - shape.Radius(),
-                 position.Y() - shape.Radius(),
-                 shape.Radius(),
-                 shape.Radius(),
+void BulletRenderer::render(Gfx &gfx, Position &position, Shape &shape, Direction &, RGB &colour) const {
+    gfx.drawRect((int) (position.X() - shape.Radius()),
+                 (int) (position.Y() - shape.Radius()),
+                 (int) shape.Radius(),
+                 (int) shape.Radius(),
                  colour);
 }
